* coders/jp2.c: Change how Jasper is intialized.  Hopefully eliminate memory leak when reading ICC color profile. * magick/utility.c (SystemCommand): Always report command error information if a command fails. * Fix some compilation warnings produced by a 64-bit build.

+++ magick/blob.c
@@ -2363,7 +2363,7 @@ MagickExport unsigned int OpenBlob(const ImageInfo *image_info,Image *image,
                   count=fread(magick,MaxTextExtent,1,image->blob->file);
                   (void) rewind(image->blob->file);
                   (void) LogMagickEvent(BlobEvent,GetMagickModule(),
-                     "  read %d magic header bytes", count);
+                     "  read %ld magic header bytes", (long) count);
 #if defined(HasZLIB)
                   if ((magick[0] == 0x1F) && (magick[1] == 0x8B) &&
                       (magick[2] == 0x08))
